src-diffuser/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
astunparse==1.6.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
docker-pycreds==0.4.0
einops==0.8.0
fastapi==0.115.5
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
GPUtil==1.4.0
grpcio==1.68.0
h11==0.14.0
h5py==3.12.1
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.0
Jinja2==3.1.4
keras==3.6.0
kornia==0.7.4
kornia_rs==0.1.7
lazy_loader==0.4
libclang==18.1.1
lightning-utilities==0.11.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
namex==0.0.8
natsort==8.4.0
networkx==3.4.2
numpy==1.26.4
omegaconf==2.3.0
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
propcache==0.2.0
protobuf==4.25.5
psutil==6.1.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pytorch-lightning==2.4.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
sentry-sdk==2.18.0
setproctitle==1.3.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.41.2
sympy==1.13.1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.1
tensorflow-intel==2.17.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
tifffile==2024.9.20
timm==1.0.11
tokenizers==0.20.3
torch==2.5.1
torchmetrics==1.6.0
torchtext==0.18.0
torchvision==0.20.1
tqdm==4.67.0
transformers==4.46.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
wandb==0.18.7
Werkzeug==3.1.3
wrapt==1.16.0
yarl==1.17.1